+++ config/routes.rb
@@ -112,6 +112,7 @@ Rails.application.routes.draw do
       get :logo
       get :request_access
       patch :set_access
+      get '(/a/:auth_token)/new_json_manifest(/*path)', action: 'new_json_manifest', as: :new_json_manifest, format: false
       get '(/a/:auth_token)/json_manifest(/*path)', action: 'json_manifest', as: :json_manifest, format: false
       get '(/a/:auth_token)/manifest(/*path)', action: 'manifest', as: :manifest, format: false
       get 'files((/a/:auth_token)(/m/:medium)/*path)', action: 'files', as: :files, format: false